CLI Plugin: Password required to log in
When opening a telnet collection the user is now asked for a password.
The password is set as a hashed value in the config file.

A general hash function has been added to the Tools-class. It may be of use
for other plugins as well (eg. the hashed backend password PR #101)