Check if Bluetooth support is available in socket module
This will check if the socket type AF_BLUETOOTH is available in the
socket module to avoid exceptions or error user will not be able
to handle. If Bluetooth support is not available hint the user to
install it and look up instructions in the readme.

See also https://github.com/smarthomeNG/smarthome/issues/229