OperationLog: Include complete log redirection to XMPP
The operation log can be used to configure a log. When logging to
a given log it is possible to configure a logger which is
additionally used when the log gets messages.

You can log the messages to a file or other handler. This example
describes logging the messages additionally though XMPP to a
XMPP contact