Database: Catch exception while rollback in dump
In case the database server is not reachable the dump will
fail and the rollback will also not work correctly. This
will keep the plugin in a wrong state (e.g. locks are still
locked and blocks further dumps or database interactions).

See also https://github.com/smarthomeNG/smarthome/issues/302