Database: Change error handling by throwing the exception in failures
Catching the exception and just ignoreing (by returning None) is
bad since the outer code can not decide if it was an error or
if there is no data. Exceptions are more precise and the outer
code can decide what to do.

Also update some logging statements, formatting, wording and
removed some debug log.

In case an error occurs in _dump() the dumped item will be
restored in the buffer list to dump it in the next cycle (this
will not lose the data).