Mailrcv: Fixed a problem that occured when an email without content-transfer-encoding was received
According to RF1341, emails can but must not contain the content-transfer-encoding header. (https://www.w3.org/Protocols/rfc1341/5_Content-Transfer-Encoding.html).
If this header does not exist, the python email package does not allow decoding of the message (at least in some versions).

To resolve this we simply set the content-transfer-encoding header to 7BIT if it does not exist.