Mailrcv Plugin: Added option to set the name of the trashfolder to use
Depending on the configuration of the IMAP server, the Trash folder can have different names.
To move a read mail to trash, the mailplugin has to know that name. Added the possibility to change
the name of the trashfolder.

If deleting the mail fails a list of abailable folders is written to the log.